requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pyrogram
tgcrypto==1.2.5
python-dotenv==1.0.1
pytz==2023.3
psycopg2-binary==2.9.9
